Rename FSDB check CLI to `fsdb_healthcheck` and switch to Click implementation
- Update `pyproject.toml` script entry: replace `fsdb_check` with `fsdb_healthcheck`.
- Modify `validation.py` to suggest the new `fsdb_healthcheck` CLI instead of a TODO comment.
- Add new CLI module `src/commons/plantdb/commons/cli/fsdb_healthcheck.py`:
  - Replace `argparse` with `click` for argument parsing.
  - Introduce `--log-level`, `--fix`, `--fix-missing`, and `--fix-extra` options.
  - Configure logger via `get_logger('fsdb_healthcheck', ...)`.
  - Implement missing‑reference fixing logic with progress bar and backup handling.
  - Stub `--fix-extra` with `NotImplementedError`.
- Remove old `fsdb_check.py` implementation.
